Fixed include paths and added function prototypes

+++ test/uut/uut.h
@@ -4,7 +4,7 @@ extern "C" {
 #endif
 #include <stdint.h>
 
-	uint32_t do_sth(uint8_t p1, uint8_t p2);
+	uint32_t do_sth(uint8_t p1, uint32_t p2);
 
 #ifdef __cplusplus
 }
